fix ushort issue

+++ backend/source/API/Domain/IdentityManagement/UserToken.cs
@@ -8,7 +8,7 @@ namespace OriolOr.Maneko.API.Domain.IdentityManagement
     public class  UserToken
     {
         public string Value;
-        public UInt16 ExpirationTime;
+        public uint ExpirationTime;
 
     } 
 }
